Move enums to ScriptUtilities for easy import in Detectors
