+++ src/Diagnostics.ModelsAndUtils/Models/Resource/App.cs
@@ -1,5 +1,5 @@
 ﻿using Diagnostics.ModelsAndUtils.Attributes;
-using Diagnostics.ModelsAndUtils.Utilities;
+using Diagnostics.ModelsAndUtils.ScriptUtilities;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -34,7 +34,8 @@ namespace Diagnostics.ModelsAndUtils.Models
         /// <summary>
         /// Resource URI
         /// </summary>
-        public string ResourceUri {
+        public string ResourceUri
+        {
             get
             {
                 return UriUtilities.BuildAzureResourceUri(SubscriptionId, ResourceGroup, Name, Provider, ResourceTypeName);
@@ -111,7 +112,7 @@ namespace Diagnostics.ModelsAndUtils.Models
         /// <returns>True, if app resource passes the filter. False otherwise</returns>
         public bool IsApplicable(IResourceFilter filter)
         {
-            if(filter is AppFilter appFilter)
+            if (filter is AppFilter appFilter)
             {
                 return ((appFilter.AppType & this.AppType) > 0) &&
                     ((appFilter.PlatformType & this.PlatformType) > 0) &&
