+++ src/Diagnostics.ModelsAndUtils/Models/ResponseExtensions/Insight.cs
@@ -1,4 +1,5 @@
-﻿using Diagnostics.ModelsAndUtils.Utilities;
+﻿using Diagnostics.ModelsAndUtils.ScriptUtilities;
+using Diagnostics.ModelsAndUtils.Utilities;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
