+++ src/Diagnostics.ModelsAndUtils/ScriptUtilities/SolutionText.cs
@@ -1,4 +1,4 @@
-﻿namespace Diagnostics.ModelsAndUtils.Utilities
+﻿namespace Diagnostics.ModelsAndUtils.ScriptUtilities
 {
     /// <summary>
     /// Signify text values for <see cref="Solution"/>s that are added in the Solution's rendering process.
